Fixed #1675: Updated exemplary streaming collector plugin
