+++ control/fixtures/fixtures.go
@@ -39,8 +39,8 @@ var (
 	PluginNameMock2 = "snap-plugin-collector-mock2"
 	PluginPathMock2 = helper.PluginFilePath(PluginNameMock2)
 
-	PluginNameStreamRand1 = "snap-plugin-stream-collector-rand1"
-	PluginPathStreamRand1 = helper.PluginFilePath(PluginNameStreamRand1)
+	PluginNameStreamingRand1 = "snap-plugin-streaming-collector-rand1"
+	PluginPathStreamingRand1 = helper.PluginFilePath(PluginNameStreamingRand1)
 
 	PluginNameMock2Grpc = "snap-plugin-collector-mock2-grpc"
 	PluginPathMock2Grpc = helper.PluginFilePath(PluginNameMock2Grpc)
