+++ glide.lock
@@ -34,7 +34,7 @@ imports:
 - name: github.com/intelsdi-x/gomit
   version: db68f6fda248706a71980abc58e969fcd63f5ea6
 - name: github.com/intelsdi-x/snap-plugin-lib-go
-  version: d1d32a057c0c23169796016876ce89634b2c012f
+  version: a8b9252d1c8305deb3c24495f1e9ce26607fcbd7
   subpackages:
   - v1/plugin
   - v1/plugin/rpc
