--- plugin/collector/snap-plugin-stream-collector-rand1/main.go
@@ -1,34 +1,0 @@
-/*
-http://www.apache.org/licenses/LICENSE-2.0.txt
-
-
-Copyright 2017 Intel Corporation
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package main
-
-import (
-	"github.com/intelsdi-x/snap-plugin-lib-go/v1/plugin"
-	"github.com/intelsdi-x/snap/plugin/collector/snap-plugin-stream-collector-rand1/rand"
-)
-
-const (
-	pluginName    = "test-rand-streamer"
-	pluginVersion = 1
-)
-
-func main() {
-	plugin.StartStreamCollector(&rand.RandCollector{}, pluginName, pluginVersion)
-}
